src/main/java/io/jaeyeon/fastcampusboard/controller/ArticleController.java
package io.jaeyeon.fastcampusboard.controller;

import io.jaeyeon.fastcampusboard.service.ArticleService;
import lombok.RequiredArgsConstructor;
import org.springframework.stereotype.Controller;
import org.springframework.ui.ModelMap;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestMapping;

import java.util.List;

/**
 * /articles/{article-id}
 * /articles/search
 * /articles/search-hashing
 */
@Controller
@RequiredArgsConstructor
@RequestMapping("/articles")
public class ArticleController {

    private final ArticleService articleService;

    @GetMapping
    public String articles(ModelMap map) {
        map.addAttribute("articles", List.of());
        return "articles/index";
    }

    @GetMapping("/{articleId}")
    public String articles(@PathVariable Long articleId, ModelMap map) {
        map.addAttribute("article", null);
        map.addAttribute("articleComment", List.of());
        return "articles/detail";
    }
}
